Removed vocab study
