--- cumulus_library/studies/vocab/.gitignore
@@ -1,1 +1,0 @@
-icd/*.parquet
