--- cumulus_library/studies/vocab/README.MD
@@ -1,6 +1,0 @@
-# Deprecation notice
-
-The vocab study is being superseded by the 
-[UMLS study](https://github.com/smart-on-fhir/cumulus-library-umls).
-If you need clinical coding definitions, use that study instead. This study will
-be removed in the next major release.
