--- cumulus_library/studies/vocab/__init__.py
@@ -1,1 +1,0 @@
-__version__ = "0.1.0"
