--- cumulus_library/studies/vocab/manifest.toml
@@ -1,17 +1,0 @@
-study_prefix = "vocab"
-
-[table_builder_config]
-file_names = [
-    "vocab_icd_builder.py",
-]
-
-[sql_config]
-file_names = [
-    "icd_legend.sql",
-]
-
-# This export is only intended for use as part of CI regression testing
-[export_config]
-export_list = [
-    "vocab__icd"
-]
