--- cumulus_library/studies/vocab/reference_sql/vocab_icd_builder.sql
@@ -1,18 +1,0 @@
--- noqa: disable=all
--- This sql was autogenerated as a reference example using the library
--- CLI. Its format is tied to the specific database it was run against,
--- and it may not be correct for all databases. Use the CLI's build 
--- option to derive the best SQL for your dataset.
-
--- ###########################################################
-
-CREATE EXTERNAL TABLE IF NOT EXISTS `cumulus_mhg_dev_db`.`vocab__icd` (
-    CUI STRING,
-    TTY STRING,
-    CODE STRING,
-    SAB STRING,
-    STR STRING
-)
-STORED AS PARQUET
-LOCATION 's3://cumulus-athena-933137588087-us-east-1/results/cumulus_user_uploads/cumulus_mhg_dev_db/vocab/icd'
-tblproperties ("parquet.compression"="SNAPPY");
