+++ tests/test_data/study_static_file/.gitignore
@@ -1,0 +1,1 @@
+icd/*.parquet
