+++ tests/test_data/study_static_file/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "0.1.0"
